mr-pixel-smith/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mr-pixel-smith"
version = "0.1.0"
description = "AI image generator powered by Ollama with Pillow watermarking"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
]

[project.scripts]
mr-pixel-smith = "mr_pixel_smith.cli:main"

[tool.setuptools.packages.find]
where = ["."]
